update port and controller

+++ backend/controllers/product.controller.js
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
-import Product from '../models/product.model';
+import Product from '../models/product.model.js';
 
 export const getProducts = async (req, res) => {
   try {
@@ -54,10 +54,16 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: 'Invalid Product Id' });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: 'Product deleted' });
   } catch (error) {
-    res.status(404).json({ success: false, message: 'Product not found' });
+    res.status(500).json({ success: false, message: 'Server error' });
   }
 };
